@@ -8,7 +8,7 @@ namespace Findus.Helpers
     {
         public static T LoadJson<T>(string jsonFilePath, JsonSerializerSettings jsonSettings = null!)
         {
-            using StreamReader r = new StreamReader(jsonFilePath);
+            using StreamReader r = new(jsonFilePath);
             string json = r.ReadToEnd();
             if (string.IsNullOrEmpty(json)) throw new ArgumentException($"Failed to parse json file {jsonFilePath}");
             return JsonConvert.DeserializeObject<T>(json, jsonSettings);
