@@ -15,7 +15,7 @@ namespace FindusWebApp.Migrations
                     AccessToken = table.Column<string>(maxLength: 1000),
                     RefreshToken = table.Column<string>(maxLength: 1000)
                 },
-                constraints: table => { table.PrimaryKey("PK_Token", x => x.RealmId); });
+                constraints: table => table.PrimaryKey("PK_Token", x => x.RealmId));
         }
 
         protected override void Down(MigrationBuilder migrationBuilder)
